requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
numpy==1.14.5
oauthlib==2.1.0
Pillow==5.1.0
pyscroll==2.17.9
PySocks==1.6.8
PyTMX==3.21.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
socketIO-client==0.7.2
tweepy==3.6.0
urllib3==1.23
websocket-client==0.48.0
wget==3.2
